REFACTOR(prometheus): Cleanup and provision mode
 ##### SUMMARY

This refactor cleans up all unused variables still present since the 18.10 release. It
also removes the use of handlers, improves the documentation, and
conditionally applies healthchecks according to enabled features.

This also introduces the now mandatory provision mode.

 ##### IMPACT

N/A

 ##### ADDITIONAL INFORMATION

All removed variables didn't have an impact on the deployment anyway
(some of them even broke it when toggled); their removal isn't
considered as having an impact on the deployment