BUGFIX/MAJOR(prometheus): Fix install on CentOS7
 ##### SUMMARY

When using the OpenIO repository and enforcing a specific version, the
install fails with 'package not found'. This fixes this behavior.

Also, the package_upstream option is removed; in the future the package
will come exclusively from OpenIO repos. For now it is only the case for
CentOS7

 ##### IMPACT
N/A

 ##### ADDITIONAL INFORMATION